Prep v231: Move elogind specific code in login/loginctl.c to login/eloginctl.c
