+++ src/login/loginctl.c
@@ -37,7 +37,7 @@
 #include "parse-util.h"
 #include "process-util.h"
 #include "signal-util.h"
-#include "spawn-polkit-agent.h"
+//#include "spawn-polkit-agent.h"
 #include "strv.h"
 #include "sysfs-show.h"
 #include "terminal-util.h"
@@ -46,45 +46,35 @@
 #include "util.h"
 #include "verbs.h"
 
-/// Additional includes for elogind
-#include "logind-action.h"
-#include "musl_missing.h"
-#include "sd-login.h"
-#include "virt.h"
+/// Additional includes needed by elogind
+#include "eloginctl.h"
 
 static char **arg_property = NULL;
 static bool arg_all = false;
 static bool arg_value = false;
 static bool arg_full = false;
 static bool arg_no_pager = false;
-static bool arg_no_wall = false;
 static bool arg_legend = true;
 static const char *arg_kill_who = NULL;
 static int arg_signal = SIGTERM;
+#if 0 /// UNNEEDED by elogind
 static BusTransport arg_transport = BUS_TRANSPORT_LOCAL;
 static char *arg_host = NULL;
-static usec_t arg_when = 0;
-static char **arg_wall = NULL;
 static bool arg_ask_password = true;
-static bool arg_ignore_inhibitors = false;
-#if 0 /// UNNEEDED by elogind
 static unsigned arg_lines = 10;
 static OutputMode arg_output = OUTPUT_SHORT;
+#else
+/// Instead we need this:
+extern BusTransport arg_transport;
+static char *arg_host = NULL;
+extern bool arg_ask_password;
+extern bool arg_no_wall;
+extern usec_t arg_when;
+extern bool arg_ignore_inhibitors;
+extern elogind_action arg_action;
 #endif // 0
-static enum action {
-        _ACTION_INVALID,
-        ACTION_HALT,
-        ACTION_POWEROFF,
-        ACTION_REBOOT,
-        ACTION_KEXEC,
-        ACTION_SUSPEND,
-        ACTION_HIBERNATE,
-        ACTION_HYBRID_SLEEP,
-        ACTION_CANCEL_SHUTDOWN,
-        _ACTION_MAX
-} arg_action;
 
-
+#if 0 /// UNNEEDED by elogind
 static void polkit_agent_open_if_enabled(void) {
 
         /* Open the polkit agent as a child process if necessary */
@@ -98,7 +88,6 @@ static void polkit_agent_open_if_enabled(void) {
         polkit_agent_open();
 }
 
-#if 0 /// UNNEEDED by elogind
 static OutputFlags get_output_flags(void) {
 
         return
@@ -1372,417 +1361,6 @@ static int terminate_seat(int argc, char *argv[], void *userdata) {
         return 0;
 }
 
-static int logind_set_wall_message(sd_bus* bus, const char* msg) {
-        _cleanup_(sd_bus_error_free) sd_bus_error error = SD_BUS_ERROR_NULL;
-        _cleanup_free_ char *m = NULL;
-        int r;
-
-        if (strv_extend(&arg_wall, msg) < 0)
-                return log_oom();
-
-        m = strv_join(arg_wall, " ");
-        if (!m)
-                return log_oom();
-
-        r = sd_bus_call_method(
-                        bus,
-                        "org.freedesktop.login1",
-                        "/org/freedesktop/login1",
-                        "org.freedesktop.login1.Manager",
-                        "SetWallMessage",
-                        &error,
-                        NULL,
-                        "sb",
-                        m,
-                        !arg_no_wall);
-
-        if (r < 0)
-                return log_warning_errno(r, "Failed to set wall message, ignoring: %s", bus_error_message(&error, r));
-
-        return 0;
-}
-
-static int elogind_cancel_shutdown(sd_bus *bus) {
-        _cleanup_(sd_bus_error_free) sd_bus_error error = SD_BUS_ERROR_NULL;
-        int r;
-
-        r = sd_bus_call_method(
-                        bus,
-                        "org.freedesktop.login1",
-                        "/org/freedesktop/login1",
-                        "org.freedesktop.login1.Manager",
-                        "CancelScheduledShutdown",
-                        &error,
-                        NULL, NULL);
-        if (r < 0)
-                return log_warning_errno(r, "Failed to talk to elogind, shutdown hasn't been cancelled: %s", bus_error_message(&error, r));
-
-        return 0;
-}
-
-static int elogind_schedule_shutdown(sd_bus *bus, enum action a) {
-        _cleanup_(sd_bus_error_free) sd_bus_error error = SD_BUS_ERROR_NULL;
-        const char *method  = NULL;
-        int r;
-
-        if (!bus)
-                return -EIO;
-
-        polkit_agent_open_if_enabled();
-
-        switch (a) {
-
-        case ACTION_POWEROFF:
-                method = "poweroff";
-                break;
-
-        case ACTION_REBOOT:
-                method = "reboot";
-                break;
-
-        default:
-                return -EINVAL;
-        }
-
-        r = logind_set_wall_message(bus, NULL);
-
-        if (r < 0) {
-                log_warning_errno(r, "Failed to set wall message, ignoring: %s",
-                                  bus_error_message(&error, r));
-                sd_bus_error_free(&error);
-        }
-
-        r = sd_bus_call_method(
-                        bus,
-                        "org.freedesktop.login1",
-                        "/org/freedesktop/login1",
-                        "org.freedesktop.login1.Manager",
-                        "ScheduleShutdown",
-                        &error,
-                        NULL,
-                        "st",
-                        method,
-                        arg_when);
-
-        if (r < 0)
-                log_error("Failed to execute operation: %s", bus_error_message(&error, r));
-
-        return r;
-}
-
-static int elogind_reboot(sd_bus *bus, enum action a) {
-        _cleanup_(sd_bus_error_free) sd_bus_error error = SD_BUS_ERROR_NULL;
-        const char *method  = NULL;
-        int r;
-        static const char *table[_ACTION_MAX] = {
-                [ACTION_REBOOT]          = "The system is going down for reboot NOW!",
-                [ACTION_POWEROFF]        = "The system is going down for power-off NOW!"
-        };
-
-        if (!bus)
-                return -EIO;
-
-        polkit_agent_open_if_enabled();
-
-        switch (a) {
-
-        case ACTION_POWEROFF:
-                method = "PowerOff";
-                break;
-
-        case ACTION_REBOOT:
-                method = "Reboot";
-                break;
-
-        case ACTION_SUSPEND:
-                method = "Suspend";
-                break;
-
-        case ACTION_HIBERNATE:
-                method = "Hibernate";
-                break;
-
-        case ACTION_HYBRID_SLEEP:
-                method = "HybridSleep";
-                break;
-
-        default:
-                return -EINVAL;
-        }
-
-        r = logind_set_wall_message(bus, table[a]);
-
-        if (r < 0) {
-                log_warning_errno(r, "Failed to set wall message, ignoring: %s",
-                                  bus_error_message(&error, r));
-                sd_bus_error_free(&error);
-        }
-
-        /* Now call elogind itself to request the operation */
-        r = sd_bus_call_method(
-                        bus,
-                        "org.freedesktop.login1",
-                        "/org/freedesktop/login1",
-                        "org.freedesktop.login1.Manager",
-                        method,
-                        &error,
-                        NULL,
-                        "b",
-                        arg_ask_password);
-
-        if (r < 0)
-                log_error("Failed to execute operation: %s", bus_error_message(&error, r));
-
-        return r;
-}
-
-static const struct {
-        HandleAction action;
-        const char*  verb;
-} action_table[_ACTION_MAX] = {
-        [ACTION_POWEROFF]     = { HANDLE_POWEROFF,     "poweroff",    },
-        [ACTION_REBOOT]       = { HANDLE_REBOOT,       "reboot",      },
-        [ACTION_SUSPEND]      = { HANDLE_SUSPEND,      "suspend",     },
-        [ACTION_HIBERNATE]    = { HANDLE_HIBERNATE,    "hibernate",   },
-        [ACTION_HYBRID_SLEEP] = { HANDLE_HYBRID_SLEEP, "hybrid-sleep" },
-};
-
-static enum action verb_to_action(const char *verb) {
-        enum action i;
-
-        for (i = _ACTION_INVALID; i < _ACTION_MAX; i++)
-                if (streq_ptr(action_table[i].verb, verb))
-                        return i;
-
-        return _ACTION_INVALID;
-}
-
-static int parse_shutdown_time_spec(const char *t, usec_t *_u) {
-        assert(t);
-        assert(_u);
-
-        if (streq(t, "now"))
-                *_u = 0;
-        else if (!strchr(t, ':')) {
-                uint64_t u;
-
-                if (safe_atou64(t, &u) < 0)
-                        return -EINVAL;
-
-                *_u = now(CLOCK_REALTIME) + USEC_PER_MINUTE * u;
-        } else {
-                char *e = NULL;
-                long hour, minute;
-                struct tm tm = {};
-                time_t s;
-                usec_t n;
-
-                errno = 0;
-                hour = strtol(t, &e, 10);
-                if (errno > 0 || *e != ':' || hour < 0 || hour > 23)
-                        return -EINVAL;
-
-                minute = strtol(e+1, &e, 10);
-                if (errno > 0 || *e != 0 || minute < 0 || minute > 59)
-                        return -EINVAL;
-
-                n = now(CLOCK_REALTIME);
-                s = (time_t) (n / USEC_PER_SEC);
-
-                assert_se(localtime_r(&s, &tm));
-
-                tm.tm_hour = (int) hour;
-                tm.tm_min = (int) minute;
-                tm.tm_sec = 0;
-
-                assert_se(s = mktime(&tm));
-
-                *_u = (usec_t) s * USEC_PER_SEC;
-
-                while (*_u <= n)
-                        *_u += USEC_PER_DAY;
-        }
-
-        return 0;
-}
-
-static int check_inhibitors(sd_bus* bus, enum action a) {
-        _cleanup_(sd_bus_message_unrefp) sd_bus_message *reply = NULL;
-        _cleanup_strv_free_ char **sessions = NULL;
-        const char *what, *who, *why, *mode;
-        uint32_t uid, pid;
-        unsigned c = 0;
-        char **s;
-        int r;
-
-        if (arg_ignore_inhibitors)
-                return 0;
-
-        if (arg_when > 0)
-                return 0;
-
-        if (geteuid() == 0)
-                return 0;
-
-        if (!on_tty())
-                return 0;
-
-        r = sd_bus_call_method(
-                        bus,
-                        "org.freedesktop.login1",
-                        "/org/freedesktop/login1",
-                        "org.freedesktop.login1.Manager",
-                        "ListInhibitors",
-                        NULL,
-                        &reply,
-                        NULL);
-        if (r < 0)
-                /* If logind is not around, then there are no inhibitors... */
-                return 0;
-
-        r = sd_bus_message_enter_container(reply, SD_BUS_TYPE_ARRAY, "(ssssuu)");
-        if (r < 0)
-                return bus_log_parse_error(r);
-
-        while ((r = sd_bus_message_read(reply, "(ssssuu)", &what, &who, &why, &mode, &uid, &pid)) > 0) {
-                _cleanup_free_ char *comm = NULL, *user = NULL;
-                _cleanup_strv_free_ char **sv = NULL;
-
-                if (!streq(mode, "block"))
-                        continue;
-
-                sv = strv_split(what, ":");
-                if (!sv)
-                        return log_oom();
-
-                if ((pid_t) pid < 0)
-                        return log_error_errno(ERANGE, "Bad PID %"PRIu32": %m", pid);
-
-                if (!strv_contains(sv,
-                                   IN_SET(a,
-                                          ACTION_HALT,
-                                          ACTION_POWEROFF,
-                                          ACTION_REBOOT,
-                                          ACTION_KEXEC) ? "shutdown" : "sleep"))
-                        continue;
-
-                get_process_comm(pid, &comm);
-                user = uid_to_name(uid);
-
-                log_warning("Operation inhibited by \"%s\" (PID "PID_FMT" \"%s\", user %s), reason is \"%s\".",
-                            who, (pid_t) pid, strna(comm), strna(user), why);
-
-                c++;
-        }
-        if (r < 0)
-                return bus_log_parse_error(r);
-
-        r = sd_bus_message_exit_container(reply);
-        if (r < 0)
-                return bus_log_parse_error(r);
-
-        /* Check for current sessions */
-        sd_get_sessions(&sessions);
-        STRV_FOREACH(s, sessions) {
-                _cleanup_free_ char *type = NULL, *tty = NULL, *seat = NULL, *user = NULL, *service = NULL, *class = NULL;
-
-                if (sd_session_get_uid(*s, &uid) < 0 || uid == getuid())
-                        continue;
-
-                if (sd_session_get_class(*s, &class) < 0 || !streq(class, "user"))
-                        continue;
-
-                if (sd_session_get_type(*s, &type) < 0 || (!streq(type, "x11") && !streq(type, "tty")))
-                        continue;
-
-                sd_session_get_tty(*s, &tty);
-                sd_session_get_seat(*s, &seat);
-                sd_session_get_service(*s, &service);
-                user = uid_to_name(uid);
-
-                log_warning("User %s is logged in on %s.", strna(user), isempty(tty) ? (isempty(seat) ? strna(service) : seat) : tty);
-                c++;
-        }
-
-        if (c <= 0)
-                return 0;
-
-        log_error("Please retry operation after closing inhibitors and logging out other users.\nAlternatively, ignore inhibitors and users with 'loginctl -i %s'.",
-                  action_table[a].verb);
-
-        return -EPERM;
-}
-
-static int start_special(int argc, char *argv[], void *userdata) {
-        sd_bus *bus = userdata;
-        enum action a;
-        int r;
-        char** wall = NULL;
-
-        assert(argv);
-
-        a = verb_to_action(argv[0]);
-
-        /* Switch to cancel shutdown, if a shutdown action was requested,
-           and the option to cancel it was set: */
-        if ( IN_SET(a, ACTION_POWEROFF, ACTION_REBOOT)
-          && (arg_action == ACTION_CANCEL_SHUTDOWN) )
-                return elogind_cancel_shutdown(bus);
-
-        r = check_inhibitors(bus, a);
-        if (r < 0)
-                return r;
-
-        /* No power off actions in chroot environments */
-        if ((a == ACTION_POWEROFF ||
-             a == ACTION_REBOOT) &&
-            (running_in_chroot() > 0) ) {
-                log_info("Running in chroot, ignoring request.");
-                return 0;
-        }
-
-        /* Check time arguments */
-        if ( IN_SET(a, ACTION_POWEROFF, ACTION_REBOOT)
-          && (argc > 1)
-          && (arg_action != ACTION_CANCEL_SHUTDOWN) ) {
-                r = parse_shutdown_time_spec(argv[1], &arg_when);
-                if (r < 0) {
-                        log_error("Failed to parse time specification: %s", argv[optind]);
-                        return r;
-                }
-        } else
-                arg_when = now(CLOCK_REALTIME) + USEC_PER_MINUTE;
-
-        /* The optional user wall message must be set */
-        if ((argc > 1) && (arg_action == ACTION_CANCEL_SHUTDOWN) )
-                /* No time argument for shutdown cancel */
-                wall = argv + 1;
-        else if (argc > 2)
-                /* We skip the time argument */
-                wall = argv + 2;
-
-        if (wall) {
-                arg_wall = strv_copy(wall);
-                if (!arg_wall)
-                        return log_oom();
-        }
-
-        /* Perform requested action */
-        if (IN_SET(a,
-                   ACTION_POWEROFF,
-                   ACTION_REBOOT,
-                   ACTION_SUSPEND,
-                   ACTION_HIBERNATE,
-                   ACTION_HYBRID_SLEEP)) {
-                if (arg_when > 0)
-                        return elogind_schedule_shutdown(bus, a);
-                else
-                        return elogind_reboot(bus, a);
-        }
-
-        return -EOPNOTSUPP;
-}
-
 static int help(int argc, char *argv[], void *userdata) {
 
         printf("%s [OPTIONS...] {COMMAND} ...\n\n"
@@ -1790,7 +1368,9 @@ static int help(int argc, char *argv[], void *userdata) {
                "  -h --help                Show this help\n"
                "     --version             Show package version\n"
                "     --no-pager            Do not pipe output into a pager\n"
+#if 1 /// elogind supports --no-wall
                "     --no-wall             Do not print any wall message\n"
+#endif // 1
                "     --no-legend           Do not show the headers and footers\n"
                "     --no-ask-password     Don't prompt for password\n"
                "  -H --host=[USER@]HOST    Operate on remote host\n"
@@ -1805,41 +1385,45 @@ static int help(int argc, char *argv[], void *userdata) {
                "  -n --lines=INTEGER       Number of journal entries to show\n"
                "  -o --output=STRING       Change journal output mode (short, short-monotonic,\n"
                "                           verbose, export, json, json-pretty, json-sse, cat)\n\n"
+#else
+                /// elogind can cancel shutdowns and allows to ignore inhibitors
+               "  -c                       Cancel a pending shutdown or reboot\n"
+               "  -i --ignore-inhibitors   When shutting down or sleeping, ignore inhibitors\n"
 #endif // 0
-               "  -c                        Cancel a pending shutdown or reboot\n"
-               "  -i --ignore-inhibitors    When shutting down or sleeping, ignore inhibitors\n"
                "Session Commands:\n"
-               "  list-sessions             List sessions\n"
-               "  session-status [ID...]    Show session status\n"
-               "  show-session [ID...]      Show properties of sessions or the manager\n"
-               "  activate [ID]             Activate a session\n"
-               "  lock-session [ID...]      Screen lock one or more sessions\n"
-               "  unlock-session [ID...]    Screen unlock one or more sessions\n"
-               "  lock-sessions             Screen lock all current sessions\n"
-               "  unlock-sessions           Screen unlock all current sessions\n"
-               "  terminate-session ID...   Terminate one or more sessions\n"
-               "  kill-session ID...        Send signal to processes of a session\n\n"
+               "  list-sessions            List sessions\n"
+               "  session-status [ID...]   Show session status\n"
+               "  show-session [ID...]     Show properties of sessions or the manager\n"
+               "  activate [ID]            Activate a session\n"
+               "  lock-session [ID...]     Screen lock one or more sessions\n"
+               "  unlock-session [ID...]   Screen unlock one or more sessions\n"
+               "  lock-sessions            Screen lock all current sessions\n"
+               "  unlock-sessions          Screen unlock all current sessions\n"
+               "  terminate-session ID...  Terminate one or more sessions\n"
+               "  kill-session ID...       Send signal to processes of a session\n\n"
                "User Commands:\n"
-               "  list-users                List users\n"
-               "  user-status [USER...]     Show user status\n"
-               "  show-user [USER...]       Show properties of users or the manager\n"
-               "  enable-linger [USER...]   Enable linger state of one or more users\n"
-               "  disable-linger [USER...]  Disable linger state of one or more users\n"
-               "  terminate-user USER...    Terminate all sessions of one or more users\n"
-               "  kill-user USER...         Send signal to processes of a user\n\n"
+               "  list-users               List users\n"
+               "  user-status [USER...]    Show user status\n"
+               "  show-user [USER...]      Show properties of users or the manager\n"
+               "  enable-linger [USER...]  Enable linger state of one or more users\n"
+               "  disable-linger [USER...] Disable linger state of one or more users\n"
+               "  terminate-user USER...   Terminate all sessions of one or more users\n"
+               "  kill-user USER...        Send signal to processes of a user\n\n"
                "Seat Commands:\n"
-               "  list-seats                List seats\n"
-               "  seat-status [NAME...]     Show seat status\n"
-               "  show-seat   [NAME...]     Show properties of seats or the manager\n"
-               "  attach NAME DEVICE...     Attach one or more devices to a seat\n"
-               "  flush-devices             Flush all device associations\n"
-               "  terminate-seat NAME...    Terminate all sessions on one or more seats\n"
+               "  list-seats               List seats\n"
+               "  seat-status [NAME...]    Show seat status\n"
+               "  show-seat [NAME...]      Show properties of seats or the manager\n"
+               "  attach NAME DEVICE...    Attach one or more devices to a seat\n"
+               "  flush-devices            Flush all device associations\n"
+               "  terminate-seat NAME...   Terminate all sessions on one or more seats\n"
+#if 1 /// elogind adds some system commands to loginctl
                "System Commands:\n"
                "  poweroff [TIME] [WALL...] Turn off the machine\n"
                "  reboot   [TIME] [WALL...] Reboot the machine\n"
                "  suspend                   Suspend the machine to memory\n"
                "  hibernate                 Suspend the machine to disk\n"
                "  hybrid-sleep              Suspend the machine to memory and disk\n"
+#endif // 1
                , program_invocation_short_name);
 
         return 0;
@@ -1851,7 +1435,9 @@ static int parse_argv(int argc, char *argv[]) {
                 ARG_VERSION = 0x100,
                 ARG_VALUE,
                 ARG_NO_PAGER,
+#if 1 /// elogind supports --no-wall
                 ARG_NO_WALL,
+#endif // 1
                 ARG_NO_LEGEND,
                 ARG_KILL_WHO,
                 ARG_NO_ASK_PASSWORD,
@@ -1865,7 +1451,9 @@ static int parse_argv(int argc, char *argv[]) {
                 { "value",           no_argument,       NULL, ARG_VALUE           },
                 { "full",            no_argument,       NULL, 'l'                 },
                 { "no-pager",        no_argument,       NULL, ARG_NO_PAGER        },
+#if 1 /// elogind supports --no-wall
                 { "no-wall",         no_argument,       NULL, ARG_NO_WALL         },
+#endif // 1
                 { "no-legend",       no_argument,       NULL, ARG_NO_LEGEND       },
                 { "kill-who",        required_argument, NULL, ARG_KILL_WHO        },
                 { "signal",          required_argument, NULL, 's'                 },
@@ -1875,8 +1463,10 @@ static int parse_argv(int argc, char *argv[]) {
 #if 0 /// UNNEEDED by elogind
                 { "lines",           required_argument, NULL, 'n'                 },
                 { "output",          required_argument, NULL, 'o'                 },
-#endif // 0
+#else
+                /// elogind allows to ignore inhibitors for system commands.
                 { "ignore-inhibitors", no_argument,     NULL, 'i'                 },
+#endif // 0
                 {}
         };
 
@@ -1885,7 +1475,11 @@ static int parse_argv(int argc, char *argv[]) {
         assert(argc >= 0);
         assert(argv);
 
-        while ((c = getopt_long(argc, argv, "hp:als:H:M:n:o:ci", options, NULL)) >= 0) {
+#if 0 /// elogind adds some system commands to loginctl
+        while ((c = getopt_long(argc, argv, "hp:als:H:M:n:o:", options, NULL)) >= 0)
+#else
+        while ((c = getopt_long(argc, argv, "hp:als:H:M:n:o:ci", options, NULL)) >= 0)
+#endif // 0
 
                 switch (c) {
 
@@ -1940,10 +1534,11 @@ static int parse_argv(int argc, char *argv[]) {
                 case ARG_NO_PAGER:
                         arg_no_pager = true;
                         break;
-
+#if 1 /// elogind supports --no-wall
                 case ARG_NO_WALL:
                         arg_no_wall = true;
                         break;
+#endif // 1
 
                 case ARG_NO_LEGEND:
                         arg_legend = false;
@@ -1974,7 +1569,7 @@ static int parse_argv(int argc, char *argv[]) {
                         arg_transport = BUS_TRANSPORT_MACHINE;
                         arg_host = optarg;
                         break;
-
+#if 1 /// elogind can cancel shutdowns and allows to ignore inhibitors
                 case 'c':
                         arg_action = ACTION_CANCEL_SHUTDOWN;
                         break;
@@ -1982,14 +1577,13 @@ static int parse_argv(int argc, char *argv[]) {
                 case 'i':
                         arg_ignore_inhibitors = true;
                         break;
-
+#endif // 1
                 case '?':
                         return -EINVAL;
 
                 default:
                         assert_not_reached("Unhandled option");
                 }
-        }
 
         return 1;
 }
@@ -2021,17 +1615,21 @@ static int loginctl_main(int argc, char *argv[], sd_bus *bus) {
                 { "attach",            3,        VERB_ANY, 0,            attach            },
                 { "flush-devices",     VERB_ANY, 1,        0,            flush_devices     },
                 { "terminate-seat",    2,        VERB_ANY, 0,            terminate_seat    },
+#if 1 /// elogind adds some system commands to loginctl
                 { "poweroff",          VERB_ANY, VERB_ANY, 0,            start_special     },
                 { "reboot",            VERB_ANY, VERB_ANY, 0,            start_special     },
                 { "suspend",           VERB_ANY, 1,        0,            start_special     },
                 { "hibernate",         VERB_ANY, 1,        0,            start_special     },
                 { "hybrid-sleep",      VERB_ANY, 1,        0,            start_special     },
                 { "cancel-shutdown",   VERB_ANY, 1,        0,            start_special     },
+#endif // 1
                 {}
         };
 
+#if 1 /// elogind can do shutdown and allows its cancellation
         if ((argc == optind) && (ACTION_CANCEL_SHUTDOWN == arg_action))
                 return elogind_cancel_shutdown(bus);
+#endif // 1
 
         return dispatch_verb(argc, argv, verbs, bus);
 }
@@ -2056,16 +1654,22 @@ int main(int argc, char *argv[]) {
         }
 
         sd_bus_set_allow_interactive_authorization(bus, arg_ask_password);
+
         r = loginctl_main(argc, argv, bus);
 
 finish:
         sd_bus_flush_close_unref(bus);
 
         pager_close();
+#if 0 /// elogind does that in elogind_cleanup()
         polkit_agent_close();
+#endif // 0
 
         strv_free(arg_property);
-        strv_free(arg_wall);
+
+#if 1 /// elogind has some own cleanups to do
+        elogind_cleanup();
+#endif // 1
 
         return r < 0 ? EXIT_FAILURE : EXIT_SUCCESS;
 }
